UpdateMore

+++ app/src/main/java/com/example/masterphone/GioHang/ThanhToan/ThanhToanActivity.java
@@ -6,10 +6,17 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.os.Bundle;
+import android.text.Editable;
+import android.text.TextUtils;
+import android.text.TextWatcher;
 import android.view.View;
 import android.widget.Button;
+import android.widget.CheckBox;
+import android.widget.CompoundButton;
+import android.widget.EditText;
 import android.widget.FrameLayout;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
@@ -17,9 +24,12 @@ import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.example.masterphone.GioHang.GioHangActivity;
+import com.example.masterphone.HomeDashboard.HomeActivity;
 import com.example.masterphone.R;
 //import com.example.masterphone.data.model.VoucherModel;
 import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
@@ -57,7 +67,12 @@ public class ThanhToanActivity extends AppCompatActivity {
     CollectionReference collectionReference;
     FirebaseUser user;
     String userID;
-
+    EditText medtDiaChi;
+//    CheckBox cbDiaChi;
+//    boolean isCheckboxChecked = false;
+//    boolean isEditTextFilled = false;
+    String diachi, ten, sodienthoai;
+    TextView btnHome;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -70,9 +85,76 @@ public class ThanhToanActivity extends AppCompatActivity {
 
         btnMuaHang = findViewById(R.id.btnPay);
         tvToTalPriceOfPay = findViewById(R.id.tvPriceOfPayFooter);
-        tvTitleVoucher = findViewById(R.id.tvTitleVoucher);
-        tvPriceOfPayContentItem4 = findViewById(R.id.tvPriceOfPayContentItem4);
+//        tvTitleVoucher = findViewById(R.id.tvTitleVoucher);
+//        tvPriceOfPayContentItem4 = findViewById(R.id.tvPriceOfPayContentItem4);
 
+        medtDiaChi = findViewById(R.id.edtDiaChi);
+        btnHome = findViewById(R.id.iconCartHome);
+        btnHome.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent i = new Intent(ThanhToanActivity.this, HomeActivity.class);
+                startActivity(i);
+            }
+        });
+//        cbDiaChi = findViewById(R.id.checkboxDiaChi);
+
+
+//        cbDiaChi.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
+//            @Override
+//            public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
+//                isCheckboxChecked = isChecked;
+//                if (isChecked) {
+//                    // Lấy tài khoản có sẵn từ Firestore và hiển thị
+//                    DocumentReference InfoProfiledocumentReference = firestore.collection("USERS").document(userID);
+//                    InfoProfiledocumentReference.get().addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
+//                        @Override
+//                        public void onSuccess(DocumentSnapshot documentSnapshot) {
+//                            String diachi = documentSnapshot.getString("Address");
+//                            medtDiaChi.setText(diachi);
+//                        }
+//                    });
+//
+//                    medtDiaChi.setEnabled(false);
+//                    medtDiaChi.setAlpha(0.5f); // Làm mờ EditText
+//                    // Thực hiện lấy tên tài khoản từ Firestore và cập nhật lên giao diện
+//                    // ...
+//                } else {
+//                    medtDiaChi.setEnabled(true);
+//                    medtDiaChi.setAlpha(1.0f); // Đặt lại trạng thái bình thường của EditText
+//                }
+//                checkPaymentButtonEnabled();
+//
+//            }
+//        });
+
+//        medtDiaChi.addTextChangedListener(new TextWatcher() {
+//            @Override
+//            public void beforeTextChanged(CharSequence s, int start, int count, int after) {
+//                // Không cần thực hiện hành động trước khi văn bản được thay đổi
+//            }
+//
+//            @Override
+//            public void onTextChanged(CharSequence s, int start, int before, int count) {
+//                // Không cần thực hiện hành động khi văn bản đang thay đổi
+//            }
+//
+//            @Override
+//            public void afterTextChanged(Editable s) {
+//                String accountName = s.toString();
+//                isEditTextFilled = !TextUtils.isEmpty(accountName);
+//                if (isEditTextFilled) {
+//                    cbDiaChi.setEnabled(false);
+//                    cbDiaChi.setAlpha(0.5f); // Làm mờ CheckBox
+//                    // Sử dụng accountName làm tên tài khoản để thanh toán
+//                    diachi = medtDiaChi.getText().toString().trim();
+//                } else {
+//                    cbDiaChi.setEnabled(true);
+//                    cbDiaChi.setAlpha(1.0f); // Đặt lại trạng thái bình thường của CheckBox
+//                }
+//                checkPaymentButtonEnabled();
+//            }
+//        });
         // load và nhận dữ liệu tổng tiền từ cart activity
         totalPrice = getIntent().getIntExtra("totalPriceFromCart", 0);
         // format giá
@@ -80,7 +162,7 @@ public class ThanhToanActivity extends AppCompatActivity {
         tvToTalPriceOfPay.setText(totalGia + " VNĐ");
 
         //load và nhận dữ liệu tổng sản phẩm
-        totalQuantity = getIntent().getIntExtra("totalQuantityFromCart", 1);
+//        totalQuantity = getIntent().getIntExtra("totalQuantityFromCart", 1);
 
         TextView tvBack = findViewById(R.id.tvBack);
         tvBack.setOnClickListener(view -> {
@@ -118,23 +200,48 @@ public class ThanhToanActivity extends AppCompatActivity {
                 });
         collectionReference = firestore.collection("USERS").document(auth.getCurrentUser().getUid()).collection("AddToCart");
 
+        DocumentReference InfoProfiledocumentReference = firestore.collection("USERS").document(userID);
+        InfoProfiledocumentReference.get().addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
+            @Override
+            public void onSuccess(DocumentSnapshot documentSnapshot) {
+                ten = documentSnapshot.getString("Fullname");
+                sodienthoai = documentSnapshot.getString("Phone");
+            }
+        });
         btnMuaHang.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                collectionReference.get().addOnCompleteListener(task -> {
-                    if (task.isSuccessful()) {
-                        if (!task.getResult().getDocuments().isEmpty()) {
-                            for (DocumentSnapshot doc : task.getResult().getDocuments()) {
-                                collectionReference.document(doc.getId()).delete();
+//                if (isCheckboxChecked || isEditTextFilled) {
+//                    // Thực hiện thanh toán
+//                    // ...
+//
+//                } else {
+//                    // Hiển thị thông báo yêu cầu nhập đầy đủ thông tin
+//                    Toast.makeText(getApplicationContext(), "Vui lòng nhập đầy đủ thông tin", Toast.LENGTH_SHORT).show();
+//                }
+                diachi = medtDiaChi.getText().toString().trim();
+                if(TextUtils.isEmpty(diachi)){
+                    Toast.makeText(getApplicationContext(), "Vui lòng nhập địa chỉ", Toast.LENGTH_SHORT).show();
+                }else if (ten.length() == 0 || sodienthoai.length() == 0){
+                    Toast.makeText(getApplicationContext(), "Vui lòng cập nhật đầy đủ thông tin", Toast.LENGTH_SHORT).show();
+                }else {
+                    collectionReference.get().addOnCompleteListener(task -> {
+                        if (task.isSuccessful()) {
+                            if (!task.getResult().getDocuments().isEmpty()) {
+                                for (DocumentSnapshot doc : task.getResult().getDocuments()) {
+                                    collectionReference.document(doc.getId()).delete();
+                                }
                             }
                         }
-                    }
-                });
-                placedOrder();
+                    });
+                    placedOrder();
+                }
             }
         });
     }
-
+//    private void checkPaymentButtonEnabled() {
+//        btnMuaHang.setEnabled(isCheckboxChecked || isEditTextFilled);
+//    }
     private void placedOrder() {
 
         // các biến lưu thời gian mua hàng
@@ -173,18 +280,23 @@ public class ThanhToanActivity extends AppCompatActivity {
         Map<String, Object> dateInfo = new HashMap<>();
         dateInfo.put("madonhang", id);
         dateInfo.put("makhachhang", userID);
-        dateInfo.put("tongSoLuong", totalQuantity);
+        dateInfo.put("ten", ten);
+        dateInfo.put("sodienthoai", sodienthoai);
+        dateInfo.put("diachi", diachi);
+//        dateInfo.put("tongSoLuong", totalQuantity);
         dateInfo.put("tongGia",totalPrice);
         dateInfo.put("ngayMua", saveCurrentDate);
         dateInfo.put("thoigianMua", saveCurrentTime);
         dateInfo.put("trangthai", trangthai);
 
+
+        String idDetail = UUID.randomUUID().toString();
         CollectionReference OrderInfocollectionReference = firestore.collection("ORDERS").document(id)
                 .collection("ORDERINFO");
-
+        DocumentReference OrderInfoDocumentReference = firestore.collection("ORDERS").document(id)
+                .collection("ORDERINFO").document(idDetail);
 
         for (ThanhToanModel model: thanhToanModelList){
-            String idDetail = UUID.randomUUID().toString();
             Map<String, Object> orderDetail = new HashMap<>();
             orderDetail.put("madonhang", id);
             orderDetail.put("machitietdonhang", idDetail);
@@ -196,7 +308,9 @@ public class ThanhToanActivity extends AppCompatActivity {
 //            orderDetail.put("ngayMua", saveCurrentDate);
 //            orderDetail.put("thoigianMua", saveCurrentTime);
 //            orderDetail.put("trangthai", trangthai);
-            OrderInfocollectionReference.add(orderDetail);
+
+//            OrderInfocollectionReference.add(orderDetail);
+            OrderInfoDocumentReference.set(orderDetail);
         }
